another_surveys/project/kafka-consumer/requirements.txt
kafka-python==2.0.2
pymongo==3.11.3
prometheus-client==0.14.1
